PR-B-3: extract ConversionClassifier (binder-side wrapper around Conversion.Classify)
Moves the binder-side conversion-emission layer out of Binder.cs into a new
internal sealed class ConversionClassifier. The pure type-pair classifier
Conversion (Conversion.cs, public sealed class with static Classify(from, to))
is unchanged and continues to be used externally; ConversionClassifier wraps
it and adds the diagnostic-emitting / BoundExpression-shaping layer.

Methods moved into ConversionClassifier (19, with previous Binder.cs lines):
- BindConversion(ExpressionSyntax, TypeSymbol, bool)                  ~15789
- BindConversion(TextLocation, BoundExpression, TypeSymbol, bool)     ~15803
- BindClrParameterConversions(...)                                    ~15992
- NeedsBindClrParameterConversion(...)                                ~16071
- BindClrMethodGroupConversion(...)                                   ~16360
- IsMethodGroupReturnCompatible(...)                                  ~16423
- BindUserMethodGroupConversion(...)                                  ~16443
- BindCallArgumentWithRefKind(...)                                     ~9594
- BindAndValidateParameterRefKind(...)                                 ~9539
- BindConditionalRefArgument(...)                                      ~9082
- InnerModifierMatchesOuter(...)                                       ~9450
- BindAndAttachParameterDefaultValue(...)                             ~10253
- TryExtractConstantDefault(...)                                      ~10305
- AppendOmittedOptionalArguments(...)                                  ~9765
- CreateOptionalDefaultArgument(...)                                   ~9794
- TryGetConstantParameterDefault(...)                                 ~10199
- TryApplyUserDefinedImplicitArgumentConversion(...)                  ~16114
- AutoDereferenceRefReturn(...)                                       ~15615
- TryBuildDisposeCall(...)                                             ~5892

Wiring:
- Binder constructs new ConversionClassifier(binderCtx, memberLookup, ...)
  in its constructor and stores it as private readonly ConversionClassifier
  conversions. All in-Binder call sites updated to conversions.X(...) or the
  static ConversionClassifier.X(...) form. No public/internal API surface
  visible to Compilation.cs, HoverComputer.cs, or tests changed.
- ConversionClassifier depends only on BinderContext (Diagnostics) and
  MemberLookup. Cross-cutting callbacks (BindExpression, IsLvalue,
  GetRefKindFromModifier, RefKindToString, IsFormattableStringTargetType,
  BindInterpolatedStringAsFormattable, CreateErasedFunctionLiteralAdapter)
  are injected as Func<> so Binder methods do not need to widen visibility.

Deliberately left on Binder:
- BuildOrderedCallArguments, ExpandParamsArguments, TryReorderArguments,
  BindCallExpression, BindConstructorCallExpression, BindExtensionFunctionCall
  — call-site / overload-resolution glue. Moves to B-4 OverloadResolver.
- BindAccessor* family — member-access dispatch. Moves to B-9 ExpressionBinder.
- CreateErasedFunctionLiteralAdapter — lambda territory. Will move to B-6
  LambdaBinder; injected as callback for now.

Behavior preservation:
- Conversion.cs untouched.
- RefactoringBaselineTests: 92 passed, 0 failed — zero hash diffs.
- Core.Tests: 2120 passed / 1 skipped (matches baseline exactly).
- Build clean (0 warnings, 0 errors).
- All GS#### diagnostics retain the same codes, locations, messages, and
  triggering conditions.

Out of scope:
- Issues #571 (lifted T -> T? conversion), #570 (slice <-> collection),
  #533, #574: architectural fixes will land in ConversionClassifier in a
  follow-up PR after the Phase-1 decomposition is complete.

LoC delta:
- Binder.cs: -1031 / +126 (net -905)
- ConversionClassifier.cs: +1148 (new file)
- Conversion.cs: unchanged

Next: PR-B-4 OverloadResolver.
